chore: Use C++17 to build etcd-cpp-api (#1073)
* Dockerfile: Use C++17 to build etcd

Also, fix #1054 where etcd event type needs to be cast to int to be
displayed.

---------

+++ src/core/nixl_listener.cpp
@@ -408,8 +408,9 @@ public:
                 std::lock_guard<std::mutex> lock(invalidated_agents_mutex);
                 invalidated_agents.push_back(agent_name);
             } else {
-                NIXL_ERROR << "Watcher for " << event.kv().key() << " received unexpected event from etcd: "
-                           << event.event_type();
+                NIXL_ERROR << "Watcher for " << event.kv().key()
+                           << " received unexpected event from etcd: "
+                           << static_cast<int>(event.event_type());
             }
         };
